Switch to DoubleMatrixMath
DoubleMatrixMath is a copy of the MatrixMath library which replaces
floats with doubles, for use on platforms where double is higher-
precision than float. (Such as the ESP8266.) Uploading that to
Github soon.

+++ src/pendulum/hotspot.ino
@@ -139,9 +139,9 @@ void updateCalibration() {
     return;
   }
   
-  float weights[elements];
-  float gravities[elements];
-  float coefficients[3];
+  double weights[elements];
+  double gravities[elements];
+  double coefficients[3];
 
   for(int i = 0; i < elements; i++) {
     if(i < 3) coefficients[i] = 0;
